Add Google OAuth to installation guide and social auth app cmd
